pyproject.toml
[project]
name = "ai-commit-reporter-groq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "groq>=1.1.1",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]
